feat: add Snakes and Ladders and Monopoly game screens
Create the first design for the monopoly game screen.
Separate screens for SnakesAndLadders and Monopoly for better customization
